Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@4815 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ GDiffView/src/nu/mine/mosher/GDiffView.java
@@ -395,8 +395,7 @@ public class GDiffView extends JFrame
         AttributeSet attr = (AttributeSet)styles.get(highlight
                 ? "highlight"
                 : "body");
-        docSrc.setCharacterAttributes(beginPoint,endPoint - beginPoint,attr,
-                true);
+        docSrc.setCharacterAttributes((int)beginPoint,(int)(endPoint - beginPoint),attr,true);
     }
 
     public static void main(String[] args) throws ClassNotFoundException, InstantiationException, IllegalAccessException, UnsupportedLookAndFeelException, InterruptedException, BadLocationException, IOException
